Fix config file as URL on Windows
Configuration for various formatter steps can be located in separate
files. Such files might live in the local filesystem and be identified
as simple paths. They can also be remote files identified by a URL.
Both types of files are copied to the output directory using Maven's
`ResourceManager` which can handle both local and remote files.

Previously, plugin tried to deduce the output file name based on the
configured string. This was done using `FileUtils#filename()` method
that expects to only work with system file separators. Issue occurred
on Windows when URL was specified. Path separator on Windows is `\`
and in URL it is `/`. Code wasn't able to extract the last path of
the URL that denotes the file name. Instead, it treated the whole URL
as a file name. URLs contain chars invalid for Windows paths and an
exception was thrown on attempt to create the file.

This commit fixes the problem by making code use predefined name
pattern for all resolved resources. Name of the output file will not
be guessed based on the configured path or URL. It will just be unique
and have 'spotless-resource-' prefix.