Require 3.1.0+ version of Maven for Maven plugin
Such runtime requirement exists because plugin uses APIs only
available starting from 3.1.0 version of Maven. They are used for
artifact resolution and come from the `org.eclipse.aether` package
which does not exist in < 3.1.0 Maven versions.

Requirement is declared in a `<prerequisites>` section of the POM.
This section is deprecated for regular projects and should only be
used by plugins to declare a minimum Maven version. Old versions will
make the plugin execution fail with an error:

```
[ERROR] Failed to execute goal c.d.s:spotless-maven-plugin:X.Y.Z:apply (default-cli) on project test: The plugin c.d.s:spotless-maven-plugin:X.Y.Z requires Maven version 3.1.0 -> [Help 1]
```

Also added a note in the README about the minimum
required Maven version.