Include Maven version requirement in the generated POM
Before this commit, it was only included in the POM used to build the
plugin. It is different from the POM that's being published to Maven Central.
Later is generated using `generatePomFileForPluginMavenPublication`
gradle task. This resulted in Maven version requirement not being
enforced and builds failing with confusing error messages.

Also added `prerequisites` section to the POM used to test the plugin.